Moved serialize form function to stand alone file to avoid duplicate code.
