+++ web/components/sections/ContactInfoSection.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 import React, { useRef, useState } from "react";
 import { classNames } from "../../functions/classNames";
 import SimpleBlockContent from "../SimpleBlockContent";
+import { serializeForm } from "../../functions/serializeForm";
 
 const classList = {
   fieldLabel: classNames("block font-medium"),
@@ -11,15 +12,6 @@ const classList = {
     "border border-gray-300",
     "focus:ring-indigo-500 focus:border-indigo-500"
   ),
-};
-
-const serializeForm = function (form) {
-  const returnObject = {};
-  const formData = new FormData(form);
-  for (let key of formData.keys()) {
-    returnObject[key] = formData.get(key);
-  }
-  return returnObject;
 };
 
 function ContactInfoSection({ contact, emailTo }) {
